Add apis for User Add filtering for Tasks
