+++ src/apps/tasks/models.py
@@ -17,6 +17,7 @@ class Task(models.Model):
     PRIORITY_CHOICES = ((1, "Low"), (2, "Medium"), (3, "High"))
     title = models.CharField(max_length=300, blank=False)  # index
     description = models.TextField(blank=True)
+    creation_date = models.DateTimeField(auto_now_add=True)
     status = models.CharField(max_length=20, choices=STATUS_CHOICES, default="pending")
     assigned_to = models.ManyToManyField(
         User, related_name="assigned_tasks", null=True, blank=True
