+++ src/apps/tasks/serializers.py
@@ -1,8 +1,6 @@
 from rest_framework import serializers
 from .models import Task
-from django.contrib.auth import get_user_model
 
-User = get_user_model()
 
 
 class TasksSerializer(serializers.ModelSerializer):
@@ -12,10 +10,3 @@ class TasksSerializer(serializers.ModelSerializer):
         # fileds = ["id" , "title" , "description"]
 
 
-class UserSerializer(serializers.ModelSerializer):
-    assigned_tasks = TasksSerializer(many=True, read_only=True)
-    created_tasks = TasksSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = User
-        fields = "__all__"
