+++ src/apps/tasks/urls.py
@@ -1,12 +1,10 @@
 from django.urls import path
 from rest_framework.routers import SimpleRouter
-from .views import TasksViewSet, UserViewSet
+from .views import TasksViewSet
 
 
 router = SimpleRouter()
 router.register('tasks' , TasksViewSet, basename='taks')
-router.register('users' , UserViewSet, basename='users')
+
 urlpatterns = [] + router.urls
-# urlpatterns = [
-# '' , include(router.urls)
-# ] 
+
