+++ src/apps/tasks/views.py
@@ -1,7 +1,8 @@
-from rest_framework import viewsets
 from .models import Task
-from .serializers import TasksSerializer, UserSerializer
+from .serializers import TasksSerializer
 from django.contrib.auth import get_user_model
+from rest_framework import viewsets
+from django_filters.rest_framework import DjangoFilterBackend
 
 User = get_user_model()
 
@@ -9,8 +10,5 @@ User = get_user_model()
 class TasksViewSet(viewsets.ModelViewSet):
     queryset = Task.objects.all()
     serializer_class = TasksSerializer
-
-
-class UserViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
+    filter_backends = [DjangoFilterBackend]
+    filterset_fields = ["title", "creation_date", "priority" , "status"]
