+++ src/apps/users/urls.py
@@ -1,0 +1,7 @@
+from rest_framework.routers import SimpleRouter
+from .views import UserViewSet
+
+
+router = SimpleRouter()
+router.register("users", UserViewSet, basename="users")
+urlpatterns = [] + router.urls
