+++ src/apps/users/views.py
@@ -1,3 +1,13 @@
-from django.shortcuts import render
+from .serializers import UserSerializer
+from django.contrib.auth import get_user_model
+from rest_framework import viewsets
+# from django_filters.rest_framework import DjangoFilterBackend
 
-# Create your views here.
+User = get_user_model()
+
+
+class UserViewSet(viewsets.ModelViewSet):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+    # filter_backends = [DjangoFilterBackend]
+    # filter_backends = ['username']
